backend/requirements.txt
openai>=1.30.0
notion-client>=2.2.1
weaviate-client>=4.6.0
pandas>=2.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
python-dotenv>=1.0.1
